Centralize check_safe_key
