+++ corehq/blobs/mixin.py
@@ -12,8 +12,12 @@ from os.path import join
 
 from corehq.blobs import BlobInfo, get_blob_db
 from corehq.blobs.exceptions import AmbiguousBlobStorageError, NotFound
-from corehq.blobs.interface import SAFENAME
-from corehq.blobs.util import ClosingContextProxy, document_method, random_url_id
+from corehq.blobs.util import (
+    ClosingContextProxy,
+    document_method,
+    random_url_id,
+    SAFENAME
+)
 from couchdbkit.exceptions import InvalidAttachment, ResourceNotFound
 from dimagi.ext.couchdbkit import (
     Document,
